requirements.txt
fastapi
uvicorn
python-multipart
pandas
requests>=2.31.0
beautifulsoup4
google-play-scraper

openai
python-dotenv


google-genai
boto3
